next function

+++ tests/compiler/test_function.py
@@ -1,4 +1,3 @@
-from ctypes import *
 from tests import eval_expr, eval
 
 def test_simple_function():
@@ -9,7 +8,7 @@ def test_simple_function():
 
         return test_function()
     """
-    assert eval(prog) == c_long(42)
+    assert eval(prog) == 42
 
 def test_args():
     prog = """\
@@ -20,7 +19,7 @@ def test_args():
         return add(10, 20)
     """
 
-    assert eval(prog) == c_long(30)
+    assert eval(prog) == 30
 
 def test_args_type_infer():
     prog = """\
@@ -30,4 +29,4 @@ def test_args_type_infer():
         return ret_byte(20), ret_float(20)
     """
 
-    assert eval(prog) == (c_byte(20), c_float(20))
+    assert eval(prog) == (20, 20)
